fix: - Money 검사 로직 추가 - bags 유니크 및 인덱스 추가 - existsByItem_Id -> existsByItem_IdAndUser_Id 변경

+++ src/main/java/xyz/catequest/spring/domain/user/controller/UserController.java
@@ -78,15 +78,15 @@ public class UserController {
 
   @PostMapping("/v1/users/money/add")
   public Response<Void> addMoney(
-      @AuthenticationPrincipal AuthUser authUser, @Positive @RequestParam Long money) {
-    userService.addMoney(authUser.getUserId(), BigInteger.valueOf(money));
+      @AuthenticationPrincipal AuthUser authUser, @Positive @RequestParam BigInteger money) {
+    userService.addMoney(authUser.getUserId(), money);
     return Response.noContent();
   }
 
   @PostMapping("/v1/users/money/minus")
   public Response<Void> minusMoney(
-      @AuthenticationPrincipal AuthUser authUser, @Positive @RequestParam Long money) {
-    userService.minusMoney(authUser.getUserId(), BigInteger.valueOf(money));
+      @AuthenticationPrincipal AuthUser authUser, @Positive @RequestParam BigInteger money) {
+    userService.minusMoney(authUser.getUserId(), money);
     return Response.noContent();
   }
 }
